+++ Changelog.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Client.Out is renamed to Client.Sinks.
 - The versioning scheme now adheres to '0.major.minor.patch' to correctly indicate impacts of new releases. (#54)
+- The 'communication' module is renamed to 'clients'.
 ### Deprecated
 ### Removed
 - communication.module() and httprequests/localserver.proxy() functions.